Move Region.copy to the base class
